Camera Update + Knife Throw in playercontroller
Moved playercontroller script and added camera control to that script. Assign the camera you want to be used for a player and the knife prefab in the editor. The camera should not be a child of the player, otherwise the camera will follow the player's rotation and movement in ways that we don't want. Do not assign a script to the camera.